Satoshi XOR proof-of-concept
As discussed on the Theory of Bitcoin interview series, it is possible to force
state propagation from tx to tx by XORing the hash of state (data) with the
key. Here I demonstrate this in the tx-experiments.js file in the simplest
possible way. I hash the value 9000 (the state, or data) and XOR that with the
address. The scripts are very similar to standard pubkeyhash type, except that
the address is hashed with data and that must be inserted into the input.