delete testnet
...and regtest, etc.

fundamentally, these alternative constants are all about displaying addresses
differently which is not helpful. it is better if applications just use
mainnet. you can send extremely low value transactions on mainnet for testing.
if you want to use testnet or regtest, that is OK too. just display mainnet
addresses and don't worry about it. this tool is for SPV which will never
require testnet or regtest.