fix yarn
I think this commit fixes yarn for people. There were two issues I recently
introduced. I renamed a bunch of packages because this was Coasian proprietary
code which I am now naming OpenSPV and licensing for the consortium. So all
packages should have "openspv" in the name and not "coasian". The second change
was that I removed yarn 3.0.2 from the repo and the default version of yarn is
1 which was incompatible with the way workspaces are being used in the project.
Both of this issues are now fixed. Theoretically, the project should work for
everyone now.