change license to OpenBSV
We are going to start using the OpenBSV license for all code to make everyone
more comfortable contributing and to keep the bureaucracy to a minimum.
However, we will also adopt an informal NDA. Do not leak any information about
the SBWC to anyone who is not a member. That includes OpenSPV source code - do
not leak the source code in any way other than bundling it inside your business
application. Anyone who violates this rule will be removed from the SBWC (but
not sued). Note that, at this time, we are publishing OpenSPV for free on npm
to make it as easy as possible to use for consortium members. But if
non-members begin to use it directly, we may remove the public access. OpenSPV
is not intended to be used directly by anyone who is not a consortium member.