'u' just means unsafe
my big number and elliptic curve code is based on the work of fedor indutny. he
prefixes a function with 'i' to mean in-place and 'u' to mean unsafe. since
these are just color for the function names, i will remove them for now and
stick with more plain names. if i ever had 'unsafe' variants, i can add back
the 'u'.