Revert "base58 WIP - i need to make a "to base" function"
This reverts commit 9187d14954d12d5f0d46b930781848c293c9e5f9.

There is a flaw with my way of handling Base58, which is any string with a
bunch of leading zeros are not convert into a number correctly because there is
no way to represent the leading zeroes. So I'd rather not bother with this -
the library I am using is fine.