invm: inverse mod
This function is necessary for ECDSA.

This code is translated from Fedor's bn.js. The tests are also taken from his
code. This is not really enough tests, but as I add the rest of the elliptic
curve code, there are a lot more tests that will be run on these methods as
they are critical for elliptic curve cryptography.

This commit also changes the behavior of mod to match bn.js, which is negative
mods should actually loop around.