MbAccount
The purpose of this new table is to replace the old keys table.

MbAccount is the record of the purchase of early access (henceforth, just
"access"). It is called MbAccount because it has all the Money Button (MB) user
account information, and MB is the only way to sign up or log in currently.