migration complete
The development and production databases have had the new tables written using
the create-mb-accounts tool. I am now going to deploy and test the app live in
production. The old table is still there if it is needed for anything. If the
app is working with no issues in production, I will proceed to prune the old
table and code.