MbPayment
The plan is to start to record the payments coming in from Money Button (MB) so
that we can properly validate them. One of the steps to validate a payment is
to validate that it is new. To know that it is new, we need to remember the old
payments. This prevents people from trying to use the same payment more than
once to pay for multiple items, i.e. double spends.