confirm payment is new
In the original Milestone 2 (Early Access), we did not check whether payments
were new or valid. This meant it would have been trivial to "hack" us by just
sending in any Money Button payment. You could use any other payment that
already existed, such as a payment to yourself.

In this update, we are now checking whether the payment is new. You can't just
use the same payment over and over to create new accounts. However, it is still
possible to use a fake payment once, because we are not yet validating the
payment. Validation will come soon.