eliminate mb_payment from mb_accounts
This column is just annoying beacause it is so much bigger than all the other
columns. It is also duplicated in the mb_payments table. So there doesn't seem
to be any reason to also have this payment data stored in the mb_accounts table
anyway.

At this time I have deleted this column from the code, but will need to
manually alter the production database to completely get rid of it.