payment is always null in mb_account
We are storing the payment in the mb_payments table, so we don't need to also
store payment here. So the payment is always null. As explained in the previous
commit, unfortunately AWS Keyspaces does not support altering a table so as to
drop a column, so we can't really get rid of this column completely without
rewriting the entire table. So we just have to leave it there, but always null,
for now.

Long-term, we shouldn't have the mb_accounts table. We will transition to email
login, which is accessible to a general audience. That is how we will
ultimately get rid of this column: by deleting the entire mb_accounts table.