DbAccount
This is, or will be, the primary database structure for user accounts. Each
user can have many accounts. Currently we track these by mbUserId, that is to
say each MB user can have many accounts.  We will abstract this soon so that a
person identified by any email, paymail, or heartmail can have any number of
user accounts.

This commit finishes the implementation of this DB structure, but we have not
yet integrated this into the application. For the moment we are still relying
on MbAccounts to be the primary account data structure.