AddressAccount
The idea here is that when an MB user signs in, I will need to query which
account they own. I can look this up by MB user ID. So for each mbUserId, I
need a table that maps those to the list of accounts they own. I was planning
to do it that way, except I want to also enable sign in with email in the
future. Thus, instead of just using mbUserId, it makes sense to use "address",
where address can be anything like [name]@[domain], e.g. a paymail, email, or
heartmail. For MB, we will use [mbUserId]@moneybutton.com, which is always a
valid paymail for that user.