AuthAddressAccount
I am working on the new account system which will need to write new accounts to
multiple tables simultaneously. Names are very important for conceptual
understanding. Suppose the user can log in with any "email address" of the form
[name]@[domain], excepe that it isn't necessarily an email address. It may be
email. It may be paymail. It may be heartmail. What do I call this? In the
context of signing into HeartMail, it makes a lot of sense to call this an
"auth address". In other words, it is an "authentication address". This is
exatly what it is. It is an address the user uses to sign into the app. And it
indexes all the accounts owned by that "user". So I have renamed the
appropriate table and variables to "authAddress" or "auth_address".