util -> dbApi
"util" is essentially an API that the front-end can use to access the database.
the front-end should not be using the DB models or data structures directly and
it should not execute queries directly. Instead, the front-end should always
use the methods inside dbApi to query the database. These methods are designed
to cover all edge cases and should not risk making some error in the DB.