auth_address -> email
Today, Money Button (MB) broke. It does not send transactions. This makes it
impossible to use the app and difficult to develop the app.

I've decided to end our dependence on MB sooner rather than later. The first
step is - forget about that auth_address name. We should just use email. In
other words, no one will be signing in by swiping MB. Users will be signing in
by inserting their email address and getting a link to click. This will work
for all users perfectly fine.

Furthermore, after this milestone is over, we will move onto the next
milestone, which will be the implementation of "stamps" where by stamps we mean
the mechanism of deposit and withdrawing stamps. Stamps are not a Bitcoin
Wallet.