email = [mbUserId]@moneybutton.com
Money Button (MB) is back online. While I don't want to depend on MB much
longer, it is simply easier for now if all users have the "email account" email
set to their mb user ID.

In other words, the user will sign in by swiping MB. When they do so, we get
their MB user ID. The paymail [mbUserId]@moneybutton.com is a stable paymail
that cannot be transferred to another user. So it uniquely identifies that
user. It makes sense to query accounts by this value. So we look up the
accounts that they have with this userId and then sign them in to which ever
one was signed into most recently.

It is easier to do it this way for now. We can then give people the option of
switching to an email address instead of MB auth in the future.

I want to point out that some users do not have an email address at all, and
that is why this option is preferrable at this time. It is the only option that
works uniformly the same for all users, thus simplifying next steps.