begin sign in code
The sign in code is partially finished but untested. Because some of the sign
in code requires many internet calls, I will need to change the dbApi to allow
me to mock the internet calls to implement more tests. I need to keep things
working in the main app while I refactor and then I will come back and test the
sign in code.