update account profile settings works. and use local fonts.
You can change your profile settings, such as name, bio, contact fee, and these
changes are reflected both in the account and also in the emailAccount as
needed. What this means is that you can see the name/bio updated in the
accounts page as well. This is non-trivial due to the way we are having to
manually maintain consistency across tables due to Cassandra/Keyspaces.

...also, I have switched back to local fonts, not hosted from Google, because I
do not see any advantages to using Google that outweigh the costs of giving
Google more of our users' data.