sign into new account upon purchase
Both the front-page and the accounts page have a button to purchase a new
account for $10.00. In either case, we now redirect to the /accounts page and
automatically sign you into the new account. Furthermore, on the /accounts
page, the new affiliate ID is whoever you were signed into last. If you are
signed in as A and make B, then the affiliate of B is A.